+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Still more comment fixes
+
 ## 0.9.3 (14 March 2026)
 
 - Fix inline comments on array elements being displaced by mutations (append,